requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.9
psycopg2==2.9.10
PyJWT==2.10.1
PyMySQL==1.1.2
python-decouple==3.8
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
